refactor: admin 폴더구조, react comfiler 적용
