+++ apps/admin/package.json
@@ -14,8 +14,8 @@
     "@tanstack/react-table": "^8.21.3",
     "axios": "^1.7.9",
     "clsx": "^2.1.1",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "react-router-dom": "^6.28.0",
     "tailwind-merge": "^2.6.0",
     "zod": "^3.24.1",
@@ -24,12 +24,13 @@
   "devDependencies": {
     "@repo/tailwind-config": "*",
     "@repo/typescript-config": "*",
-    "@types/react": "^18.3.12",
-    "@types/react-dom": "^18.3.1",
+    "@types/react": "^19.0.0",
+    "@types/react-dom": "^19.0.0",
     "@typescript-eslint/eslint-plugin": "^8.19.1",
     "@typescript-eslint/parser": "^8.19.1",
-    "@vitejs/plugin-react-swc": "^3.5.0",
+    "@vitejs/plugin-react": "^5.1.2",
     "autoprefixer": "^10.4.17",
+    "babel-plugin-react-compiler": "^1.0.0",
     "eslint": "^9.0.0",
     "eslint-plugin-react": "^7.37.2",
     "postcss": "^8.4.35",
